+++ app/Http/Controllers/CustomerTypeController.php
@@ -56,6 +56,7 @@ class CustomerTypeController extends Controller
         privilegeLevel(self::config['privilege'], CAN_CRUD);
 
         $data = [
+            'id'    => null,
             'name'  => null,
             'color' => 'primary',
         ];
@@ -264,7 +265,7 @@ class CustomerTypeController extends Controller
         ];
     }
 
-    public function validateInput($request, $id = null)
+    public function validateInput(Request $request, $id = null)
     {
         // validasi
         $rules = [
@@ -280,31 +281,15 @@ class CustomerTypeController extends Controller
 
         $validator = Validator::make($request->all(), $rules, $messages);
 
-        if($validator->fails()){
-            return redirect()->back()->withErrors($validator)->withInput($request->all)->send();
-        } else {
-            return true;
+        // jika hanya validate input
+        if ($request->validate) {
+            return $validator->errors();
+        }else{
+            if($validator->fails()){
+                return redirect()->back()->withErrors($validator)->withInput($request->all)->send();
+            } else {
+                return true;
+            }
         }
-    }
-
-    public function validateForm(Request $request) {
-        // dapatkan id
-        $id = $request->id ?: null;
-
-        // validasi
-        $rules = [
-            'name'                  => 'required|min:3|max:100|unique:customer_types,name,'.$id,
-        ];
-
-        $messages = [
-            'name.required'         => 'Nama wajib diisi',
-            'name.min'              => 'Nama minimal 3 karakter',
-            'name.max'              => 'Nama maksimal 100 karakter',
-            'name.unique'           => 'Nama sudah terdaftar',
-        ];
-
-        $validator = Validator::make($request->all(), $rules, $messages);
-
-        return $validator->errors();
     }
 }