+++ resources/views/layout/user/privilege/input.blade.php
@@ -20,9 +20,10 @@
         <div id="main" class="col-md-9">
             <div class="card mb-3">
                 <div class="card-body">
+                    <input type="hidden" name="id" value="{{ old('id', $data->id) }}">
                     <div class="mb-3">
                         <label class="form-label">Name<span class="text-danger">*</span></label>
-                        <input type="text" name="name" class="form-control w-50 @error('name') is-invalid @enderror" value="{{ old('name', $data->name) }}" placeholder="Nama Hak Akses / Privilege">
+                        <input type="text" name="name" class="form-control w-50 @error('name') is-invalid @enderror" value="{{ old('name', $data->name) }}" placeholder="Nama Hak Akses / Privilege" validate>
                         @error('name')
                             <div class="invalid-feedback">
                                 {{ $message }}
@@ -57,6 +58,9 @@
                     <div class="mb-3">
                         @include('form.privilege', ['name' => 'jobs'])
                     </div>
+                    <div class="mb-3">
+                        @include('form.privilege', ['name' => 'misc'])
+                    </div>
                     <hr class="border-dashed">
                     <div class="mb-3">
                         @include('form.color')
@@ -71,4 +75,10 @@
         </div>
     </div>
     @include('form.header.end')
+@endsection
+
+@section('script')
+    <script>
+        initInput('{{ $config['url'] }}');
+    </script>
 @endsection