+++ resources/views/layout/warranty/table.blade.php
@@ -4,12 +4,12 @@
         <thead>
         <tr>
             @include('component.table.title', ['title' => '#', 'column' => 'warranties.id', 'sortable' => true, 'class' => 'text-center'])
+            @include('component.table.title', ['title' => 'Customer', 'column' => 'customers.name', 'sortable' => true])
             @include('component.table.title', ['title' => 'Type', 'column' => 'warranties.type', 'sortable' => true])
             @include('component.table.title', ['title' => 'Model', 'column' => 'warranties.model', 'sortable' => true])
             @include('component.table.title', ['title' => 'Serial', 'column' => 'warranties.serial', 'sortable' => true])
             @include('component.table.title', ['title' => 'No. Warranty', 'column' => 'warranties.warranty_no', 'sortable' => true])
             @include('component.table.title', ['title' => 'Purchase Date', 'column' => 'warranties.purchase_date', 'sortable' => true])
-            @include('component.table.title', ['title' => 'Customer', 'column' => 'customers.name', 'sortable' => true])
             {{-- Jika can CRUD maka munculkan tombol--}}
             @if(getUserLevel($config['privilege']) >= CAN_CRUD)
                 @include('component.table.title', ['title' => 'Action', 'column' => 'action', 'sortable'=> false, 'class' => 'text-center align-middle'])
@@ -32,6 +32,20 @@
                 <tr>
                     <td class="text-center text-muted w-1-slot">{{ $table['column'] == 'id' && $table['sort'] == 'desc' ? $data->total() - ($data->firstItem() + $key) + 1 : $data->firstItem() + $key }}</td>
                     <td>
+                        @if($row->customer)
+                            <span>{{ ucwords($row->customers->name) }}</span>
+                            <br><span class="text-muted">{{ $row->customers->phone }}</span>
+                            @if($row->customers->phone2)
+                                <br><span class="text-muted">{{ $row->customers->phone2 }}</span>
+                            @endif
+                            @if($row->customers->phone3)
+                                <br><span class="text-muted">{{ $row->customers->phone3 }}</span>
+                            @endif
+                        @else
+                            <span class="text-muted">-</span>
+                        @endif
+                    </td>
+                    <td>
                         {{ $row->type == 0 ? 'Out' : 'In' }}
                     </td>
                     <td>{{ $row->model }}</td>
@@ -43,17 +57,9 @@
                             <span class="text-muted">-</span>
                         @endif
                     </td>
-                    <td>{{ date('d/m/Y', strtotime($row->purchase_date)) }}</td>
                     <td>
-                        @if($row->customer)
-                            <span>{{ ucwords($row->customers->name) }}</span>
-                            <br><span class="text-muted">{{ $row->customers->phone }}</span>
-                            @if($row->customers->phone2)
-                                <br><span class="text-muted">{{ $row->customers->phone2 }}</span>
-                            @endif
-                            @if($row->customers->phone3)
-                                <br><span class="text-muted">{{ $row->customers->phone3 }}</span>
-                            @endif
+                        @if($row->type == IN_WARRANTY)
+                            <span>{{ date('d/m/Y', strtotime($row->purchase_date)) }}</span>
                         @else
                             <span class="text-muted">-</span>
                         @endif