+++ resources/views/sidebar/quick.blade.php
@@ -1,1 +1,4 @@
-<a href="#" class="btn btn-primary w-100 text-white">@svg('heroicon-s-clipboard-document-list', 'icon') New Ticket</a>
+<a href="#" class="btn btn-success w-100 text-white mb-2">@svg('heroicon-s-user', 'icon') New Customer</a>
+<a href="#" class="btn btn-warning w-100 text-white mb-2">@svg('heroicon-s-credit-card', 'icon') New Warranty</a>
+<a href="#" class="btn btn-primary w-100 text-white mb-2">@svg('heroicon-s-clipboard-document-list', 'icon') New Ticket</a>
+<a href="#" class="btn btn-secondary w-100 text-white">@svg('heroicon-s-briefcase', 'icon') New Job</a>